@@ -11,22 +11,24 @@
     "@mui/icons-material": "^6.4.1",
     "@mui/material": "^6.4.1",
     "@svgr/webpack": "^8.1.0",
+    "@tanstack/react-query": "^5.86.0",
     "@testing-library/react": "^16.2.0",
     "cra-template-typescript": "1.2.0",
+    "dayjs": "^1.11.18",
     "framer-motion": "^12.0.5",
     "moment": "^2.30.1",
+    "notistack": "^3.0.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-router-dom": "^7.1.3",
-    "react-scripts": "5.0.1",
     "web-vitals": "^4.2.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "docs": "typedoc --out ../docs/frontend src"
+    "dev": "vite",
+    "build": "tsc --noEmit && vite build",
+    "preview": "vite preview",
+    "docs": "typedoc --out ../docs/frontend src",
+    "typecheck": "tsc --noEmit"
   },
   "eslintConfig": {
     "extends": [
@@ -48,6 +50,20 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.6.3",
-    "typescript": "^4.9.5"
+    "@types/node": "^24.3.0",
+    "@types/react": "^18.3.24",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^5.0.2",
+    "typescript": "^5.9.2",
+    "vite": "^7.1.4",
+    "vite-tsconfig-paths": "^5.1.4"
+  },
+  "pnpm": {
+    "overrides": {
+      "react": "^18.3.1",
+      "react-dom": "^18.3.1",
+      "@types/react": "^18.3.3",
+      "@types/react-dom": "^18.3.0"
+    }
   }
 }
